nvofapi-vk: Renames
